requirements.txt
colorama
openpyxl
python-docx
python-pptx
PyYAML